TMS_8.5/python-sample/requirements.txt
gevent
protobuf
grpcio
